chore: Extract test decorators to doctest/parts/decorators.h
